Fixing sf download links
Chromium, free-cad, mumble, open-office, seashore, and unetbootin
were all using url's that ended in download rather than the
filename. This has been changed, sha's are preserved.

Mumble does not have a 'latest' download alias in the new url
scheme. It does however specifiy a version so the Cask was updated
to reflect this.

Closes #229

Signed-off-by: phinze <paul.t.hinze@gmail.com>