Upgrade LyX to v2.0.7.1, Mac only bugfix release
http://www.lyx.org/News
Bugfix Release for OSX Only: LyX 2.0.7.1 released.

February 1, 2014

We are pleased to announce the release of LyX 2.0.7.1. This is a
bugfix release that affects Mac OSX only. Users of other operating
systems do not need to update to this version, and no binaries will be
provided for other operating systems.

There were some issues with the menus in LyX 2.0.7 on OSX.
Specifically, "dynamic" menu items were lost, and there was a problem
restoring the menus from full-screen operation. Both issues are fixed
in 2.0.7.1.