Update hear.rb
Version update today to 1.2.1.  I grabbed the file via the s3 bucket and then ran `shasum -a 256` on it.  Not sure if this is the right procedure, so I wanted to try this and get some feedback.  

But what is the `appcast` url and sha sum there fore?  Obviously I didn't update those as I wasn't sure what is going on there.