Add Angband.app v3.5.1
Updated as requested to parse the major/minor/release versioning
from the version string and use them to populate the download link.

Scheme handles Angband as versioned from 3.3 onward for download
links.  As for version numbering, it handles anything the various
maintainers have created since switching to the major/minor/release
versioning scheme with 2.5.x.  This includes appending letters with
the numbers.

(I however did not manage to find a 'frog-knows' version on rephial
to test, but it's positively ancient, and should work anyway...)