audio-hijack 4.4.0
The newer version of Audio Hijack that require macOS 14.4 is
`AudioHijack-ARK.zip` instead of `AudioHijack.zip` (the older
variant). This updates the cask to properly update to 4.4.0 while
also handling the older version.

This also reverts the `livecheck` block to the `Sparkle` approach,
since we need to gate the check to specific macOS versions and trying
to accomplish that by parsing the Support Center page may be too
challenging/unreliable. The `system` value in the `Sparkle` feed URLs
will need to be updated periodically, of course.