zettelkasten: downgrade to 3.2022.8
`zettelkasten` was upgraded to 3.2023.10 in #158445 but the related
release has since been removed from the GitHub project, so the cask
`url` gives a 404 (Not Found) response. The repository contains tags
for v3.2023.10 (2023-10-24) and v3.2023.12 (2023-12-30) but the latest
release (with the `Package.dmg.zip` file we use) is 3.2022.8.

This downgrades the cask to the latest available release, 3.2022.8.