threema-work: adjust version to 1.2.41
The `Threema Work.app` in the `Threema-Work-Latest.dmg` file is
version 1.2.41 and the 2.0 versions on the Changelog page
(https://threema.ch/en/work/changelog) are all beta versions at the
moment. This aligns the cask `version` with the actual app version.

This also updates the `livecheck` block to resolve the URL redirection
and adjusts the regex to make sure it only matches stable versions.