geogebra: update livecheck
This updates the `livecheck` block to move the inline regex in the
`strategy` block into a `#regex` call (passing the value into the
block).

Since the `location` value only contains a URL, this also loosens the
regex to reduce the chance it will fail to match a version if the
filename format changes in the future. We can always tighten it up
again in the future if the looseness leads to an incorrect match.