font-poppins-latin: add livecheck
By default, livecheck checks the Git tags for `font-poppins-latin` but
it returns 4.003 instead of 5.001, as there is no 5.001 tag. The cask
zip file is found in the `products` directory of the repository and
the 5.001 version comes from the filename.

This adds a `livecheck` block that checks the `products` directory
page and matches versions from the related zip file links.