ideamaker: update livecheck
The existing `livecheck` block for `ideamaker` is giving an `Unable
to get versions` error, as the existing download page no longer
links to dmg files. Instead, the download page links to separate
pages for Apple Silicon and Intel.

This addresses the issue by updating the `livecheck` block to account
for this new setup.