stella: add livecheck
By default, livecheck checks the Git tags for `stella` but upstream
tagged `7.0a` and `7.0b` versions after 7.0 that appear to only apply
to Windows, so livecheck is reporting 7.0b as newest. We use the dmg
release asset in the cask and that's still 7.0.

This adds a `livecheck` block that matches the version from the
dmg file in the "latest" release on GitHub. Since the tag version and
filename version can differ, I've set up the `livecheck` block and
cask `url` to be able to handle a mismatch.