triage: fix `extract_plist livecheck` label
This label was never being applied because the regex only matched `strategy :extract_plist` when indented by 2 spaces, this is always at least 4 spaces because it is nested inside of a `livecheck do` block. The newline check shouldn't be necessary regardless, and removing it causes this to also match in the case that the `livecheck` is nested within another block, ie. `on_arm do`.