miniconda: update livecheck
The existing `livecheck` block for `miniconda` contains a `#strategy`
call but it doesn't specify a strategy symbol. We expect the symbol to
always be provided when `#strategy` is used but we enforce this using
a RuboCop and the `livecheck` block RuboCops don't apply to casks yet
(I'm currently working on this).

This adds the missing `#strategy` symbol and renames the `strategy`
block argument to use the standard `page` name for `PageMatch`.