acronis-true-image 41592
Acronis True Image wasn't deprecated, it was just temporarily
rebranded as Acronis Cyber Protect Home Office. Now it's back!

TODO: define zap procedure.
At the moment, the official uninstaller is an app on the distribution
image (DMG). It needs to be somehow preserved at the installation time.
However, such an uninstaller matching the current installed version
is not guaranteed as the user can upgrade Acronis True Image via a
built-in upgrade feature.