openframeworks: update livecheck
The "latest" release on GitHub for `openframeworks` is a `latest` tag,
which doesn't correspond to a version. We could check all recent
releases using the `GithubReleases` strategy but this updates the
`livecheck` block to check the first-party download page instead, as
it links to the newest file for macOS on GitHub and requires less
data transfer.