cloudflare-warp: update uninstall
A user reported that some `cloudflare-warp` symlinks remained in
`/usr/local/bin` after uninstallation. This adds `uninstall` args to
remove the current symlinks.

Besides that, this adds the app bundle's `uninstall.sh` script to the
uninstall steps, as that handles some additional things that we can't
address in the cask (e.g., deleting keychain secrets).