libreoffice-still: downgrade to 24.2.7
This reverts `libreoffice-still` to 24.2.7, as that's what is listed
as the "Previous Version" on the "Release Notes" page. It seems that
upstream uploaded a newer version to the `stable` directory before
removing it, so the existing `livecheck` block picked it up at the
time. Currently the `stable` directory only contains 24.8.4 and
24.2.7, as expected.

This updates the `livecheck` block to check the "Release Notes" page,
which lists both the latest and previous ("still") version. This
regex is naturally more fragile but it works for now.