leader-key 1.10.1
The appcast URL in the `livecheck` block for `leader-key` is currently
giving a 404 (Not Found) response. Checking the source files, it
appears that upstream is serving the appcast from github.io now, so
this updates the `livecheck` block accordingly. The appcast links to
the app.zip file found in the GitHub release assets, so this also
switches the cask `url` to that source while updating to the newest
version, 1.10.1.