shearwater-cloud: update livecheck
The existing `livecheck` block for `shearwater-cloud` is giving an
`Unable to get versions` error, as the download page no longer links
to files. Instead, upstream has updated the download buttons for
macOS and Windows to point to an intermediary page that uses
JavaScript to fetch a JSON file containing filenames and redirect to
the relevant file. This updates the `livecheck` block to match the
version information from the dmg filename in the JSON file.