gologin: update livecheck
The existing `livecheck` block for `gologin` is failing with the
following error: `curl: (56) Unrecognized content encoding type.
libcurl understands deflate, gzip content encodings.` This occurs
because livecheck uses the curl `--compressed` option by default and
the server response contains a `Content-Encoding: aws-chunked`
header that curl doesn't know how to handle. The check works fine if
the `--compressed` option isn't used but we currently don't have a
way of unsetting that option in livecheck, so we have to work around
this for now.

This updates the `livecheck` block to check the first-party download
page and match the version from the release notes link. This isn't
ideal but it will work for the time being.