dolphin@dev 2503a-444
This updates `dolphin@dev` to the latest development version,
2503a-444. The cask had been stuck on an older version because the
regex wasn't able to match versions due to the letter at the end of
the first part (2503a). This updates the regex to be able to match
this format.

Besides that, this updates the macOS dependency to resolve the related
audit failure ("Upstream defined :big_sur as the minimum OS version
but the cask declared :catalina").