psychopy: update version and url
The 2025.1.1 `psychopy` release on GitHub no longer provides the file
that the cask currently references, so the download will predictably
fail. The current dmg file uses a similar filename as some previous
releases (albeit, upstream is inconsistent about the filename format
from one release to the next), where there is no second part of the
cask `version` and it uses a `-py3.10` suffix instead of `_3.10`.
This updates the `version` and `url` accordingly.