processing: update livecheck
I worked up a modified `livecheck` block for `processing` yesterday
morning to address `GithubLatest` not working for the latest release
but didn't create a PR and Klaus beat me to it in #212817. Remarkably,
we produced almost exactly the same `livecheck` block, so this simply
incorporates the differences (i.e., adding an explanatory comment and
using `#filter_map` instead of `#map`).