hdfview: update livecheck
The existing `livecheck` block for `hdfview` uses the `GithubLatest`
strategy but it's currently returning an "Unable to get versions"
error, as the "latest" release on GitHub is the `snapshots` release
instead of `v3.3.2`. We could use the `GithubReleases` strategy to
identify the latest stable release but it would also require a
somewhat involved `strategy` block, as upstream doesn't always
reliably mark unstable releases as pre-release (e.g., `HDFView-3.3.99`
isn't marked as pre-release, so livecheck returns it as the newest
version when using the default `GithubReleases` strategy logic).

This addresses the issue by updating the `livecheck` block to check
the homepage and match the version from the filename text (the
download links are gated behind a login). This isn't ideal, as it
doesn't align with the cask URL source (GitHub releases) and the
hdfgroup.org server doesn't have compression enabled, so we download
the full 180 KB each time. Hopefully this is only a temporary
workaround but there are enough issues with how upstream handles
GitHub releases and unstable versions that this may be the only
reliable approach for now.