creative: update livecheck
The existing `livecheck` block regex for `creative` contains an
unusual `(\d+(?:\d)+)+` pattern that's intended to match the build
number from text like "Creative 1.5.3 build 103.zip". This is flagged
by code scanning as an inefficient regular expression and `\d+` would
be sufficient here. I imagine this pattern was intended to be
`(\d+(?:\.\d+)*)`, which is what we typically use when matching a
number without dots (i.e., we allow for optional dots, so the regex
doesn't break if upstream adds dots in the future).

This updates the regex accordingly, which should resolve the code
scanning alert. Besides that, I've updated instances of `\s` to `\s+`,
so the regex doesn't fail to match if the text contains more than one
space at any point.