gisto: add livecheck
By default, livecheck checks the Git tags for `gisto` and it's
returning 2.2.6 as newest but it's been two days since the tag was
created and there's no release on GitHub yet. This adds a `livecheck`
block that uses the `GithubLatest` strategy, as there can be a notable
gap between the tag and release.