openboardview 9.95.2
`openboardview` is autobumped but the workflow is failing to update
to version 9.95.2 because the dmg filename now includes an `-x86_64`
suffix. Besides that, this also fails `brew audit` because the cask
needs a Rosetta caveat and the app is unsigned. This updates the
version, adds a Rosetta caveat, and adds a `disable!` call using the
date that we've been using for this situation.