soulseek 2024-6-30
`soulseek` is autobumped but the workflow failed to update to version
2024-6-30 because the download files are now hosted on Backblaze B2.
This updates the version and URL accordingly.

Besides that, this tweaks the `livecheck` block regex to use a more
standard pattern, which will also continue to match if the `href`
attribute only contains the filename instead of being absolute. In
this particular context, we don't need to strictly delimit the start
of the filename (i.e., using `/`) to accurately match the dmg version.