joplin: downgrade to 3.3.13
The 3.4.x releases of Joplin are currently all marked as pre-release
on GitHub. We updated the cask to 3.4.6 when it was the "latest"
release but it apparently has some bugs, so this downgrades the cask
to 3.3.13 for the time being.

Besides that, I've updated the `livecheck` block to check the
first-party download page, in hopes that this page is only updated
to link to a new dmg when the version is properly stable. There's no
guarantee that it will avoid this particular release issue but it's
worth a shot.