Update older cask renames to use current name
We have an existing `"ankiapp-anki": "ankiapp"` rename in
`cask_renames.json` but we recently added an `"ankiapp": "algoapp"`
rename. We assumed that the chained renames would be handled correctly
(i.e., `ankiapp-anki` -> `ankiapp` -> `algoapp`) but that may not be
true for all `brew` commands. From a quick test, `brew info` simply
outputs a "Cask ankiapp-anki was renamed to ankiapp" warning, so it
only handles the first rename. I haven't checked whether an upgrade
would be handled correctly but, considering that at least one `brew`
command doesn't handle chained renames, there's maybe something to be
said for updating the existing rename to point to the current cask
instead.

We also have existing renames for `"clinq": "sipgate-clinq"` and
`"sipgate-clinq": "sipgate"`, so this also updates the older `clinq`
rename accordingly.