font-lxgw-neoxihei: add livecheck
By default, livecheck uses the `Git` strategy to check for new
versions of `font-lxgw-neoxihei` but this is currently returning a
version that's marked as pre-release on GitHub. The release uses a
tag with the stable version format, so the `Git` strategy isn't able
to reliably identify the newest stable version in this situation.
This addresses the issue by adding a `livecheck` block using the
`GithubLatest` strategy.