zappy 4.9.0
`zappy` is autobumped but the workflow failed to update to 4.9.0
because the delimiter in the file name changed back from an underscore
to the usual hyphen (it appears to have been a one-off issue for
4.8.9). Besides that, this also fails `brew audit` with an "Upstream
defined :sequoia as the minimum macOS version but the cask declared a
depends_on stanza with a minimum macOS version of :big_sur" error.
This updates the version, `url`, and `depends_on macos:` value
accordingly.