postico@1: update livecheck
The existing `livecheck` block for `postico@1` is producing an
`Unable to get versions` error, as the download URL on the 1.x
homepage redirects to the download page for 2.x, which also links to
files for 1.x. This updates the `livecheck` block to check the current
download page, updating the regex and `strategy` block for the
different content.

For what it's worth, we specifically have to match versions with 3+
parts, as there's also text like "Postico 1.5" that precedes the full
version text and the regex will match that instead (leading to an
incomplete version). Alternatively, we could match the [full] version
text within a `li` element (as the incomplete version text is in a
`div`) but that may not hold up as well over time.