renamer 7.0.1
`renamer` is autobumped but the workflow failed to update to 7.0.1
because livecheck is returning an `Unable to get versions` error, as
the Renamer 6 appcast now returns a 404 (Not Found) response. The new
app checks the `latest-mac.yml` file from the latest GitHub release
in the incbee/renamer-7 repository and that doesn't seem to exist
(at least not publicly) but there is a renamer-7-releases repository,
so this updates the `livecheck` block to check the file from those
releases.