taskade: update livecheck
The existing `livecheck` block for `taskade` is producing an `Unable
to get versions` error, presumably because the HTML for the downloads
page is a 6.44 MB download (_with_ compression) and curl can time out
before the server finishes sending the response body. The download URL
for macOS isn't part of the main download page content and is instead
found in the footer that's part of every page. This resolves the issue
by updating the `livecheck` block to check the homepage, which is ~100
KB compressed and hopefully won't run into the same issue over time.