wetype 1.4.2
`wetype` is autobumped but the existing `livecheck` block is producing
an `Unable to get versions` error because the file name in the Mac
download link doesn't include version information (it's
`WeTypeInstaller_1001.zip` instead of the expected
`WeTypeInstaller_1.4.2_529.zip`). Other Mac-related URLs include the
version in their file name, so this adds some fallback logic to the
`livecheck` block's `strategy` block to try to match the version from
the `InstallInfo` URL. The installer download URL works with this
version, so this approach works for this particular version/scenario
but hopefully upstream simply returns to the usual file name format in
future versions. Until then, this will allow us to update to version
1.4.2 and we can always remove this fallback logic if it causes
problems in the future.