qlc+ 5.0.0
`qlc+` is autobumped but the workflow failed to update to version
5.0.0 because it's only for Apple Silicon (as mentioned in the
upstream release notes). This updates the cask accordingly, splitting
the versions based on architecture and maintaining the last Intel
release as a legacy version for now. It's technically possible to
continue checking for new Intel versions on the download page but
upstream doesn't appear to create new releases for older major
versions after a newer major version is released, so it would likely
just be a waste of effort.