atuin-desktop 0.1.5
`atuin-desktop` is autobumped but the workflow failed to update to
version 0.1.5 because livecheck checked the Git tags but the related
release hadn't been created yet (it took ~3 hours after the tag was
pushed). This updates the version and modifies the `livecheck` block
to use the `GithubLatest` strategy, as there can be a notable gap
between tag and release.