kindle-comic-converter 9.2.1
`kindle-comic-converter` is autobumped but the workflow failed to
update to version 9.2.1 because `brew audit` failed with an "Artifact
defined :big_sur as the minimum macOS version but the cask declared a
depends_on stanza with a minimum macOS version of :monterey" error.
This updates the version and splits the `depends_on macos:` value
for ARM and Intel, as the Intel requirement is macOS 10.13 and the
ARM requirement is macOS 11.