gcs 5.42.0
`gcs` is autobumped but the workflow failed to update to version
5.42.0 because `brew audit` failed with an "Artifact defined :big_sur
as the minimum macOS version but the cask declared no minimum macOS
version" error in reference to the ARM app. The Intel version has a
minimum macOS requirement of Catalina (10.15), so it can rely on the
implicit macOS dependency but it's necessary to have an explicit
declaration as `brew style` will fail if we only have the `on_arm`
`depends_on macos:` value. This updates the version and adds
appropriate `depends_on macos:` values.