libreoffice-still 25.2.7
`libreoffice-still` is autobumped but the workflow failed to update
to version 25.2.7 because `brew audit` failed with an "Artifact
defined :big_sur as the minimum macOS version but the cask declared
no minimum macOS version" error for the ARM app. The Intel app has a
macOS dependency of Catalina (10.15) or later and it can rely on the
implicit macOS dependency but we have to include an explicit
`depends_on macos:` value, as `brew style` will fail if we only have
a `depends_on macos:` value in the `on_arm` block. This updates the
version and adds appropriate `depends_on macos:` values.