blender 4.5.4
`blender` is autobumped but the workflow failed to update to version
4.5.4 because livecheck is returning an "Unable to get versions" error
as www.blender.org is behind Cloudflare and is now inaccessible due
to protections. The website is only accessible after going through a
Cloudflare turnstile check in a browser and the existing check is
broken regardless of IP address, user agent, etc.

This updates the version and modifies the `livecheck` block to match
versions from the directory listing page where the release assets are
found. This could theoretically result in livecheck surfacing a
version before the release is announced but hopefully upstream is
careful about their release process, as this is one of few available
sources of version information left.