elektron-transfer: disable
`elektron-transfer` is autobumped but the livecheck block has been
consistently producing an "Unable to get versions" error for some
time. The `livecheck` block fails because the website is behind
Cloudflare and the protections prevent us from fetching pages outside
of a browser context (i.e., this fails regardless of user agent, IP
address, etc.).

Unfortunately, the dmg file is also served from the website and is
subject to the same Cloudflare protections, so this cask is
effectively unusable. This removes the `livecheck` block and disables
the cask.