mythic 0.5.0
`mythic` is autobumped but the workflow failed to update to version
0.5.0 because livecheck failed with an "Unable to get versions" error.
This check works locally but seems to consistently fail in the
autobump workflow as well as in CI tests. We're not able to autobump
this as a result, so this adds a `no_autobump!` call to the cask.

This manually updates the cask to the newest version, adjusting the
`url` to use the new file URL format referenced in the appcast. This
is the first version using this URL format, so it's not yet clear if
the hexadecimal identifier in the path is specific to version 0.5.0 or
if it's a constant and this is an unversioned URL. We'll find out with
the next version but I've set up the cask for now with the assumption
that this identifier will change in the future. If it doesn't change,
then we can inline the identifier in the `url`, set the `sha256` to
`:no_check`, and revert this `livecheck` block change at that time.