luanti: update livecheck
The existing `livecheck` block for `luanti` is producing an "Unable
to get versions" error because the downloads page only contains
minimal HTML to redirect browsers to /en/downloads/. The server
response doesn't contain a `Location` header to redirect, so curl
isn't able to follow the redirection. This fixes the check by updating
the `livecheck` block URL to resolve the redirection.