insomnia@alpha 12.0.0
`insomnia@alpha` is autobumped but the workflow failed to update to
version 12.0.0 because the livecheck audit failed for this version.
The API endpoint that the `livecheck` block checks only returns a JSON
response if the provided version in the URL is lower than the newest
version. We work around this by using `#{version.major}.0.0` in the
`livecheck` block URL and this works for versions that don't end in
".0.0". However, this predictably fails for a new major version like
12.0.0, preventing autobump from being able to create a version bump
PR due to the `livecheck` audit failure.

We can't resolve this issue by falling back to the cask `version` in
the `strategy` block because `Json.find_versions` contains a guard
that will return early if the response body is empty (as `Json.parse`
will error if it's given an empty string), so the `strategy` block
isn't executed in that scenario. This resolves the issue by using the
previous major version in the `livecheck` block URL if the cask
version contains ".0.0". This is slightly different behavior than the
`livecheck` block for `insomnia`, as the version may have an unstable
suffix in this context. This approach works for 12.0.0 and hopefully
it will continue to work in the future, so autobump will also handle
new major versions.