unison-app 2.53.8
`unison-app` is autobumped but the workflow failed to update to
version 2.53.8 because the arch suffix for Intel is `x86-64` for this
release instead of the previous `x86_64`. This updates the version
and adjusts the related `arch` value accordingly. Besides that, the
app fails Gatekeeper checks, so this adds a `disable!` call with the
future date we've been using for this scenario.