ricochet-refresh: add Intel macOS dependency
The ARM app for `ricochet-refresh` depends on macOS 11 or later and we
had to add that dependency to resolve a `brew audit` error. However,
the Intel app has a macOS requirement of 10.14 or later, so it uses
the current implicit macOS requirement of `>= :catalina`. `brew style`
will fail if we only have a `depends_on macos:` value in an `on_arm`
block, so we have to make the implicit macOS dependency explicit for
Intel.