tinkerwell 5.6.0
`tinkerwell` is autobumped but the workflow failed to update to
version 5.6.0 because `brew audit` failed with a "Cask is deprecated
because it failed Gatekeeper checks but all artifacts now pass!"
error. I manually verified that the app passes Gatekeeper checks, so
this updates the version and removes the `disable!` call.