mouseless@preview 0.5.0-preview.1
`mouseless@preview` is autobumped but the workflow failed to update
to version 0.5.0-preview.1 because it is marked as pre-release on
GitHub. This updates the version and adds the cask to
`github_prerelease_allowlist.json` using an `any` value, as the cask
will also update to [stable] releases that aren't marked as
pre-release.