Rename eid-viewer to beid-viewer
`eid-viewer` was merged before `beid-token` but these are related and
should use the same naming convention. We decided on the `beid-`
prefix in the `beid-token` PR and I suggested that `eid-viewer` should
be renamed and updated to reflect feedback from that PR but this did
not happen. This renames `eid-viewer` to `beid-viewer` to align it
with `beid-token`.